Add feature to configure emails post send.
After sending the email we can add the email toe DB as the class name and allow
the end user to change the subject, who will be receiving it and who it sends
too.

Will need to add features to allow certain properties to not be changeable via
the admin. Like making to to email send to the user like a confirmation.