Change arguments if passed through template.
If arguments are passed through to the init and we havnt been created this will
make it so those arguments are used for this email if not we will use the ones
on self.

Should be used when you want to change the from_email to be the user that sent
the email or if you want to the to to be a users email address.