Allow test users, enabling/disabling emails.
We need the ability to disable/enable emails. Along with allowing people to
test the output and make sure its coming to them rather than the default
receivers.

This can be used for when a AE/PM/or Dev wants to make sure the email is being
put out correctly and want to test forms that send emails.